Add files via upload
We know that with the tail recursive, as well as the iterative dynamic programming
version of the Fibonacci sequence algorithm, we do not have to calculate the same values
repeatedly to get the right answer. This alone means that it must have a better runtime than
the recursive iteration we implemented back in lab 2. Let’s look at the algorithms and check this
prediction.
To start, let’s see how many times the tail recursive version will run. This function is
almost like a for loop. Starting at n, we execute some operations and decrease n before the
start of the next call. Because these operations are simply setting variables, they can be done in
constant time, so we can ignore these. The base case, or when the loop ends is when n=0,
meaning, this function will run n times giving it a linear runtime complexity of O(n). Now, the
original recursive implementation is a little less straight forward. It too starts with n; however, it
basically has two recursive calls at the end. It has to calculate ALL of the previous values of not
only n-1, but n-2 as well, which creates a lot of redundancy. We can create a complex
recurrence relation for it; however, we know that an exponential running time is when the
calculations double as input grows. By drawing a tree of the calls, we see this to be true; f(3)
would run 1 additional time without including base case, input f(4) would run 3 additional times
not including base case, input f(5) would run 6 additional times not including base cases, and so
on. Therefore, it will have a runtime of O(2n), which is significantly slower than linear time.
Running a timer on the two algorithms with the same input sizes, I got 3.5ms on average for the
tail recursive implementation and 136.6ms on average for the original recursive
implementation, which makes sense. Interestingly enough, we can also see that with increasing
inputs, the tail recursive version is much faster. It can calculate input 1000 in the blink of an eye
while the other has trouble calculating 40 or so.